edit transactions

+++ src/actions/transaction.actions.ts
@@ -3,7 +3,7 @@
 import db from "@/server/db";
 import { accountTable, transactionTable } from "@/server/db/schema";
 import { validateRequest } from "@/server/lucia";
-import { AddTransactionSchema } from "@/types";
+import { AddTransactionSchema, TransactionSchema } from "@/types";
 import { desc, eq } from "drizzle-orm";
 import { z } from "zod";
 
@@ -103,7 +103,108 @@ export const getTransactions = async () => {
   }
 };
 
-export const deleteTransaction = async (id: number) => {
+export const editTransaction = async ({
+  values,
+  ogAmount,
+  ogType,
+}: {
+  values: z.infer<typeof TransactionSchema>;
+  ogAmount: number;
+  ogType: string;
+}) => {
+  const { user } = await validateRequest();
+  if (!user) {
+    return {
+      error: "You must be logged in to add an account",
+    };
+  }
+
+  var accountIdNum: number = +values.accountId;
+
+  const account = await db.query.accountTable.findFirst({
+    where: (table) => eq(table.id, accountIdNum),
+  });
+
+  if (!account) {
+    return {
+      error: "Account not found",
+    };
+  }
+
+  try {
+    const updatedTransaction = await db
+      .update(transactionTable)
+      .set({
+        type: values.type,
+        amount: values.amount,
+        category: values.category,
+        account: account?.name,
+        accountId: accountIdNum,
+        dateTime: values.dateTime,
+        note: values.note,
+      })
+      .where(eq(transactionTable.id, values.id))
+      .returning({
+        id: transactionTable.id,
+        userId: transactionTable.userId,
+
+        type: transactionTable.type,
+        amount: transactionTable.amount,
+        category: transactionTable.category,
+        account: transactionTable.account,
+        accountId: transactionTable.accountId,
+        dateTime: transactionTable.dateTime,
+        note: transactionTable.note,
+      });
+
+    let newBalance = account.balance;
+    if (ogType !== values.type && ogAmount !== values.amount) {
+      if (values.type == "Expense") {
+        newBalance = newBalance - (ogAmount + values.amount);
+      } else if (values.type == "Income") {
+        newBalance = newBalance + (ogAmount + values.amount);
+      }
+    } else if (ogType !== values.type && ogAmount === values.amount) {
+      if (values.type == "Expense") {
+        newBalance -= 2 * values.amount;
+      } else if (values.type == "Income") {
+        newBalance += 2 * values.amount;
+      }
+    } else if (ogType === values.type && ogAmount !== values.amount) {
+      if (values.amount > ogAmount) {
+        if (values.type == "Expense")
+          newBalance = account.balance - (values.amount - ogAmount);
+        else if (values.type == "Income")
+          newBalance = account.balance + (values.amount - ogAmount);
+      } else {
+        if (values.type == "Expense")
+          newBalance = account.balance + (ogAmount - values.amount);
+        else if (values.type == "Income")
+          newBalance = account.balance - (ogAmount - values.amount);
+      }
+    }
+    await db
+      .update(accountTable)
+      .set({ balance: newBalance })
+      .where(eq(accountTable.id, accountIdNum));
+
+    return {
+      data: updatedTransaction,
+      success: "Transaction updated successfully",
+    };
+  } catch (error: any) {
+    return {
+      error: error.message,
+    };
+  }
+};
+
+export const deleteTransaction = async (
+  id: number,
+  accountId: number,
+  type: string,
+  amount: number
+) => {
   const { user } = await validateRequest();
 
   if (!user) {
@@ -112,8 +213,31 @@ export const deleteTransaction = async (id: number) => {
     };
   }
 
+  const account = await db.query.accountTable.findFirst({
+    where: (table) => eq(table.id, accountId),
+  });
+
+  if (!account) {
+    return {
+      error: "Account not found",
+    };
+  }
+
   try {
+    let balance = account.balance;
+    if (type === "Expense") {
+      balance += amount;
+    } else if (type === "Income") {
+      balance -= amount;
+    }
+
     await db.delete(transactionTable).where(eq(transactionTable.id, id));
+    await db
+      .update(accountTable)
+      .set({
+        balance,
+      })
+      .where(eq(accountTable.id, accountId));
 
     return {
       success: "Transaction deleted successfully",
